requirements.txt
crewai[tools]
requests
python-dotenv
streamlit